feat(backend): create new models
