+++ apps/backend/src/entities/collection.rs
@@ -14,8 +14,6 @@ pub struct Model {
     pub last_updated_on: DateTimeUtc,
     pub name: String,
     pub description: Option<String>,
-    #[graphql(skip)]
-    pub created_by_user_id: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
@@ -24,14 +22,8 @@ pub enum Relation {
     CollectionToEntity,
     #[sea_orm(has_many = "super::review::Entity")]
     Review,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::CreatedByUserId",
-        to = "super::user::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    User,
+    #[sea_orm(has_many = "super::user_to_collection::Entity")]
+    UserToCollection,
 }
 
 impl Related<super::collection_to_entity::Entity> for Entity {
@@ -46,9 +38,9 @@ impl Related<super::review::Entity> for Entity {
     }
 }
 
-impl Related<super::user::Entity> for Entity {
+impl Related<super::user_to_collection::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::User.def()
+        Relation::UserToCollection.def()
     }
 }
