+++ apps/backend/src/entities/mod.rs
@@ -1,4 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
 
 pub mod prelude;
 
@@ -19,5 +19,6 @@ pub mod review;
 pub mod seen;
 pub mod user;
 pub mod user_measurement;
+pub mod user_to_collection;
 pub mod user_to_entity;
 pub mod workout;
