+++ apps/backend/src/entities/prelude.rs
@@ -1,4 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
 
 pub use super::calendar_event::Entity as CalendarEvent;
 pub use super::collection::Entity as Collection;
@@ -17,5 +17,6 @@ pub use super::review::Entity as Review;
 pub use super::seen::Entity as Seen;
 pub use super::user::Entity as User;
 pub use super::user_measurement::Entity as UserMeasurement;
+pub use super::user_to_collection::Entity as UserToCollection;
 pub use super::user_to_entity::Entity as UserToEntity;
 pub use super::workout::Entity as Workout;
